feat: [utils] split packages
