+++ src/utils/packages/async/README.md
@@ -1,0 +1,20 @@
+# async
+
+Utility functions collection.
+
+## Installation
+
+```bash
+npm install @ohkit/async
+```
+
+## Usage
+
+```typescript
+import { PromisePipe } from '@ohkit/async';
+```
+
+## Development
+
+- Run tests: `npm test`
+- Build: `npm run build`
