+++ src/utils/packages/async/src/__tests__/index.test.ts
@@ -1,0 +1,8 @@
+import { describe, expect, it } from '@jest/globals';
+import { PromisePipe } from '../index';
+
+describe('utils', () => {
+  it('should export utilities', () => {
+    expect(PromisePipe).toBeDefined();
+  });
+});
