+++ src/utils/packages/async/src/index.ts
@@ -1,1 +1,2 @@
+// Utils index file
 export * from './promise-pipe';
