+++ src/utils/packages/auto-reporter/.npmignore
@@ -1,0 +1,7 @@
+__tests__/
+*.stories.*
+*.test.*
+plopfile.js
+tsconfig.json
+./node_modules
+.npmrc
