+++ src/utils/packages/auto-reporter/README.md
@@ -1,0 +1,20 @@
+# auto-reporter
+
+Utility functions collection.
+
+## Installation
+
+```bash
+npm install @ohkit/auto-reporter
+```
+
+## Usage
+
+```typescript
+import { ... } from '@ohkit/auto-reporter';
+```
+
+## Development
+
+- Run tests: `npm test`
+- Build: `npm run build`
