+++ src/utils/packages/auto-reporter/src/__tests__/index.test.ts
@@ -1,0 +1,8 @@
+import { describe, expect, it } from '@jest/globals';
+import { auto-reporter } from '../index';
+
+describe('utils', () => {
+  it('should export utilities', () => {
+    expect(auto-reporter).toBeDefined();
+  });
+});
