+++ src/utils/packages/auto-reporter/src/index.ts
@@ -1,0 +1,2 @@
+// Utils index file
+export * from './auto-reporter';
