+++ src/utils/packages/auto-reporter/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src"],
+  "exclude": ["dist", "node_modules"]
+}
