+++ src/utils/packages/platform/README.md
@@ -1,0 +1,20 @@
+# platform
+
+Utility functions collection.
+
+## Installation
+
+```bash
+npm install @ohkit/platform
+```
+
+## Usage
+
+```typescript
+import { ... } from '@ohkit/platform';
+```
+
+## Development
+
+- Run tests: `npm test`
+- Build: `npm run build`
