+++ src/utils/packages/prefix-classname/README.md
@@ -1,0 +1,20 @@
+# prefix-classname
+
+Utility functions collection.
+
+## Installation
+
+```bash
+npm install @ohkit/prefix-classname
+```
+
+## Usage
+
+```typescript
+import { ... } from '@ohkit/prefix-classname';
+```
+
+## Development
+
+- Run tests: `npm test`
+- Build: `npm run build`
