+++ src/utils/packages/prefix-classname/src/__tests__/index.test.ts
@@ -1,0 +1,8 @@
+import { describe, expect, it } from '@jest/globals';
+import { prefixClassname } from '../index';
+
+describe('utils', () => {
+  it('should export utilities', () => {
+    expect(prefixClassname).toBeDefined();
+  });
+});
