+++ src/utils/packages/prefix-classname/src/index.ts
@@ -1,0 +1,2 @@
+// Utils index file
+export * from './prefix-classname';
