+++ src/utils/packages/react-helper/.npmrc
@@ -1,0 +1,2 @@
+registry=https://registry.npmmirror.com/
+# registry=https://registry.npmjs.org/
