+++ src/utils/packages/react-helper/README.md
@@ -1,0 +1,20 @@
+# react-helper
+
+Utility functions collection.
+
+## Installation
+
+```bash
+npm install @ohkit/react-helper
+```
+
+## Usage
+
+```typescript
+import { ... } from '@ohkit/react-helper';
+```
+
+## Development
+
+- Run tests: `npm test`
+- Build: `npm run build`
