+++ src/utils/packages/react-helper/src/react-helper.ts
@@ -1,0 +1,4 @@
+
+export const reactHelper = () => {
+    return 'xxx';
+};
