+++ src/utils/src/index.ts
@@ -1,5 +1,5 @@
-export * from './prefix-classname';
-export * from './react-helper';
-export * from './platform';
-export * from './async';
-export * from './auto-reporter';
+export * from '@ohkit/async';
+export * from '@ohkit/auto-reporter';
+export * from '@ohkit/react-helper';
+export * from '@ohkit/platform';
+export * from '@ohkit/prefix-classname';
