--- src/utils/src/react-helper/index.ts
@@ -1,2 +1,0 @@
-export * from './assign-ref';
-export * from './use';
